Implement base collection logic
* Start to implement blockhitinformation to return information about hit
* Define hit method in blockinterface and hit method in iteminterface to define the informations that we use on hit logic
* Change CollectInformation to VolumeState and remove inheritance logic because is not needed
* Start to implement a hand item to have a empty item during test
* Add BlockCollectionService logic and handle call in InteractionComponent

+++ OctoAwesome/OctoAwesome/IInventoryableDefinition.cs
@@ -14,7 +14,7 @@ namespace OctoAwesome
         /// <summary>
         /// Gibt das Volumen für eine Einheit an.
         /// </summary>
-        decimal VolumePerUnit { get; }
+        int VolumePerUnit { get; }
 
         /// <summary>
         /// Gibt an, wie viele dieses Items im Inventar in einem Slot gestapelt werden können (in Anzahl der Blöcke, nicht in Vielfachen der <see cref="VolumePerUnit"/>!).
